+++ Users/Tasks/Sensor/tsk_sensor.h
@@ -11,7 +11,8 @@
 #include "Device/IMU/dvc_imu.h"
 
 // 通道值，1表示轨迹
-extern float Channels[8];
+extern uint8_t Channels[8];
+
 // 偏航角，偏航加速度
 extern float Yaw;
 extern float GyroZ;
